apps/ml/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
scikit-learn
